Add folder refactoring and css styling
