+++ internal/logging/logger.go
@@ -3,6 +3,7 @@ package logging
 import (
 	"log/slog"
 	"os"
+	"testing"
 )
 
 // Logger instance
@@ -11,17 +12,37 @@ var Logger *slog.Logger
 // InitLogger initializes the logger
 func InitLogger() {
 	if Logger != nil {
-		return // Avoid re-initialization
+		return
+	}
+	// Check if running in test mode
+	if isTestMode() {
+		Logger = slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{
+			Level: slog.LevelWarn,
+		}))
+		slog.SetDefault(Logger)
+		Logger.Warn("Logging disabled in test mode")
+		return
 	}
 
-	// Create logs directory if not exists
-	os.Mkdir("logs/app.log", 0755)
+	// Ensure logs directory exists
+	logDir := "logs"
+	if _, err := os.Stat(logDir); os.IsNotExist(err) {
+		err := os.Mkdir(logDir, 0755)
+		if err != nil {
+			panic("Failed to create logs directory: " + err.Error())
+		}
+	}
 
-	// Clear previous logs
+	// Define log file path
 	logFilePath := "logs/app.log"
-	os.WriteFile(logFilePath, []byte{}, 0644)
 
-	// Open log file
+	// Clear previous logs (overwrite with an empty file)
+	err := os.WriteFile(logFilePath, []byte{}, 0644)
+	if err != nil {
+		panic("Failed to clear log file: " + err.Error())
+	}
+
+	// Open log file (Create if not exists)
 	logFile, err := os.OpenFile(logFilePath, os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0644)
 	if err != nil {
 		panic("Failed to open log file: " + err.Error())
@@ -36,4 +57,9 @@ func InitLogger() {
 	slog.SetDefault(Logger)
 
 	Logger.Info("Logger initialized")
+}
+
+// Function to detect if we are running in test mode
+func isTestMode() bool {
+	return testing.Testing()
 }
