+++ web/templates/index.html
@@ -1,6 +1,8 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
+    <link rel="stylesheet" href="/static/css/styles.css">
+
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Web Page Analyzer</title>
@@ -11,7 +13,7 @@
 
 <form action="/analyze" method="post">
     <label for="url">Enter URL:</label>
-    <input type="text" id="url" name="url" required>
+    <input type="text" id="url" name="url">
     <button type="submit">Analyze</button>
 </form>
